# migration: Update case for multifd and postcopy migration