According to bug 2169733, multifd and postcopy migration will fail on RHEL 8. So update some parameters for postcopy cases.