@cliping 2023-08-09
Before on RHEL 8:
TestFail: error: internal error: unable to execute QEMU command 'migrate-set-capabilities': Postcopy is not yet compatible with multifd

After on RHEL 8:
 (1/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_tcp_transport.with_memtune.with_postcopy.p2p: PASS (225.66 s)
 (2/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_tcp_transport.with_memtune.with_postcopy.non_p2p: PASS (224.67 s)
 (3/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_tcp_transport.without_memtune.with_postcopy.p2p: PASS (224.00 s)
 (4/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_tcp_transport.without_memtune.with_postcopy.non_p2p: PASS (224.11 s)
 (5/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_auto_converge.with_memtune.with_postcopy.p2p: PASS (223.45 s)
 (6/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_auto_converge.with_memtune.with_postcopy.non_p2p: PASS (224.29 s)
 (7/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_auto_converge.without_memtune.with_postcopy.p2p: PASS (224.72 s)
 (8/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_auto_converge.without_memtune.with_postcopy.non_p2p: PASS (223.74 s)

 (1/4) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy_abort.with_memtune.with_postcopy.p2p: PASS (241.81 s)
 (2/4) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy_abort.with_memtune.with_postcopy.non_p2p: PASS (240.53 s)
 (3/4) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy_abort.without_memtune.with_postcopy.p2p: PASS (240.96 s)
 (4/4) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy_abort.without_memtune.with_postcopy.non_p2p: PASS (240.72 s)

Test result on RHEL 9:
 (1/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_tcp_transport.with_memtune.with_postcopy.p2p: PASS (278.64 s)
 (2/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_tcp_transport.with_memtune.with_postcopy.non_p2p: PASS (264.43 s)
 (3/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_tcp_transport.without_memtune.with_postcopy.p2p: PASS (256.88 s)
 (4/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_tcp_transport.without_memtune.with_postcopy.non_p2p: PASS (256.95 s)
 (5/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_auto_converge.with_memtune.with_postcopy.p2p: PASS (257.85 s)
 (6/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_auto_converge.with_memtune.with_postcopy.non_p2p: PASS (258.04 s)
 (7/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_auto_converge.without_memtune.with_postcopy.p2p: PASS (257.31 s)
 (8/8) type_specific.io-github-autotest-libvirt.migration.migration_performance_tuning.migration_zerocopy.zerocopy_and_parallel_and_auto_converge.without_memtune.with_postcopy.non_p2p: PASS (258.27 s)